Minor updates to help compiles
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@8863 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/os2/app.cpp
@@ -764,7 +764,7 @@ bool wxApp::ProcessMessage(
     // We must relay Timer events to wxTimer's processing function
     //
     if (pMsg->msg == WM_TIMER)
-        wxTimerProc(NULL, 0, pMsg->mp1, 0);
+        wxTimerProc(NULL, 0, (int)pMsg->mp1, 0);
 
     //
     // For some composite controls (like a combobox), wndThis might be NULL
